Split users into 3 separate files
