+++ user/db.ts
@@ -1,139 +1,19 @@
-import express, { Response } from "express";
+import { z } from "zod";
 import {
   SQLExecutable,
+  queryFirst,
   queryResults,
   hasResults,
-  queryFirst,
-} from "./dbconnection";
-import { ServerEnvironment } from "./env";
-import { withValidatedRequest } from "./validation";
-import { z } from "zod";
-import { Result } from "./utils";
-import { Connection } from "@planetscale/database";
+} from "../dbconnection";
+import { Result } from "../utils";
+import {
+  DEFAULT_USER_SETTINGS,
+  User,
+  UserSettings,
+  UserToProfileRelationStatus,
+} from "./models";
 
-export const userNotFoundBody = (userId: string) => ({
-  userId,
-  error: "user-not-found",
-});
-
-export const userNotFoundResponse = (res: Response, userId: string) => {
-  res.status(404).json(userNotFoundBody(userId));
-};
-
-/**
- * Creates routes related to user operations.
- *
- * @param environment see {@link ServerEnvironment}.
- * @returns a router for user related operations.
- */
-export const createUserRouter = (environment: ServerEnvironment) => {
-  const router = express.Router();
-
-  router.post("/", async (req, res) => {
-    await withValidatedRequest(req, res, CreateUserSchema, async (data) => {
-      const result = await environment.conn.transaction(async (tx) => {
-        const registerReq = Object.assign(data.body, { id: res.locals.selfId });
-        return await registerNewUser(tx, registerReq);
-      });
-
-      if (result.status === "error") {
-        return res.status(400).json({ error: result.value });
-      }
-      return res.status(201).json(result.value);
-    });
-  });
-
-  router.post("/friend/:userId", async (req, res) => {
-    const result = await environment.conn.transaction(async (tx) => {
-      return await sendFriendRequest(tx, res.locals.selfId, req.params.userId);
-    });
-    return res
-      .status(result.statusChanged ? 201 : 200)
-      .json({ status: result.status });
-  });
-
-  router.get("/self", async (_, res) => {
-    const user = await userWithId(environment.conn, res.locals.selfId);
-    if (!user) {
-      return userNotFoundResponse(res, res.locals.selfId);
-    }
-    return res.status(200).json(user);
-  });
-
-  router.get("/self/settings", async (_, res) => {
-    const settings = await environment.conn.transaction(async (tx) => {
-      return await userSettingsWithId(tx, res.locals.selfId);
-    });
-    if (settings.status === "error") {
-      return userNotFoundResponse(res, res.locals.selfId);
-    }
-    return res.status(200).json(settings.value ?? DEFAULT_USER_SETTINGS);
-  });
-
-  router.patch("/self/settings", async (req, res) => {
-    await withValidatedRequest(
-      req,
-      res,
-      PatchUserSettingsRequestSchema,
-      async (data) => {
-        const result = await environment.conn.transaction(async (tx) => {
-          return await overwriteUserSettings(tx, res.locals.selfId, data.body);
-        });
-        if (result.status === "error") {
-          return userNotFoundResponse(res, res.locals.selfId);
-        }
-        return res.status(204).send();
-      }
-    );
-  });
-
-  router.patch("/self", async (req, res) => {
-    await environment.conn.transaction(async (tx) => {
-      const result = await updateSelf(tx, {
-        selfId: res.locals.selfId,
-        ...req.body,
-      });
-      console.log("result:" + result);
-      return res.status(200).json({ result });
-    });
-  });
-  return router;
-};
-
-/**
- * A zod schema for {@link UserSettingsSchema}.
- */
-export const UserSettingsSchema = z.object({
-  isAnalyticsEnabled: z.boolean(),
-  isCrashReportingEnabled: z.boolean(),
-  isEventNotificationsEnabled: z.boolean(),
-  isMentionsNotificationsEnabled: z.boolean(),
-  isChatNotificationsEnabled: z.boolean(),
-  isFriendRequestNotificationsEnabled: z.boolean(),
-});
-
-/**
- * A type representing a user's settings.
- */
-export type UserSettings = z.infer<typeof UserSettingsSchema>;
-
-/**
- * The default user settings, which enables all fields.
- */
-export const DEFAULT_USER_SETTINGS = {
-  isAnalyticsEnabled: true,
-  isCrashReportingEnabled: true,
-  isEventNotificationsEnabled: true,
-  isMentionsNotificationsEnabled: true,
-  isChatNotificationsEnabled: true,
-  isFriendRequestNotificationsEnabled: true,
-} as const;
-
-const PatchUserSettingsRequestSchema = z.object({
-  body: UserSettingsSchema.partial(),
-});
-
-const overwriteUserSettings = async (
+export const overwriteUserSettings = async (
   conn: SQLExecutable,
   userId: string,
   settings: Partial<UserSettings>
@@ -208,7 +88,7 @@ const insertUserSettings = async (
   );
 };
 
-const userSettingsWithId = async (
+export const userSettingsWithId = async (
   conn: SQLExecutable,
   id: string
 ): Promise<Result<UserSettings | undefined, "user-not-found">> => {
@@ -236,35 +116,13 @@ const queryUserSettings = async (conn: SQLExecutable, userId: string) => {
   );
 };
 
-/**
- * A type representing the main user fields.
- */
-export type User = {
-  id: string;
-  name: string;
-  handle: string;
-  bio?: string;
-  profileImageURL?: string;
-  creationDate: Date;
-  updatedAt?: Date;
-};
-
-const userWithId = async (conn: SQLExecutable, userId: string) => {
+export const userWithId = async (conn: SQLExecutable, userId: string) => {
   return await queryFirst<User>(conn, "SELECT * FROM user WHERE id = :userId", {
     userId,
   });
 };
 
-/**
- * A type representing the relationship status between 2 users.
- */
-export type UserToProfileRelationStatus =
-  | "not-friends"
-  | "friend-request-pending"
-  | "friends"
-  | "blocked";
-
-const sendFriendRequest = async (
+export const sendFriendRequest = async (
   conn: SQLExecutable,
   senderId: string,
   receiverId: string
@@ -321,26 +179,6 @@ const twoWayUserRelation = async (
   };
 };
 
-export type userUpdateRequest = {
-  selfId: string;
-  name: string;
-  bio: string;
-  handle: string;
-};
-
-export const updateSelf = async (
-  conn: SQLExecutable,
-  request: userUpdateRequest
-) => {
-  await conn.execute(
-    `
-    UPDATE user 
-    SET name = :name, bio = :bio, handle = :handle
-    WHERE id = :selfId 
-  `,
-    request
-  );
-};
 const makeFriends = async (conn: SQLExecutable, id1: string, id2: string) => {
   await conn.execute(
     `
@@ -363,14 +201,28 @@ const addPendingFriendRequest = async (
   );
 };
 
-const CreateUserSchema = z.object({
-  body: z.object({
-    name: z.string().max(50),
-    handle: z.string().regex(/^[a-z_0-9]{1,15}$/),
-  }),
-});
+export type userUpdateRequest = {
+  selfId: string;
+  name: string;
+  bio: string;
+  handle: string;
+};
 
-const registerNewUser = async (
+export const updateSelf = async (
+  conn: SQLExecutable,
+  request: userUpdateRequest
+) => {
+  await conn.execute(
+    `
+    UPDATE user 
+    SET name = :name, bio = :bio, handle = :handle
+    WHERE id = :selfId 
+  `,
+    request
+  );
+};
+
+export const registerNewUser = async (
   conn: SQLExecutable,
   request: RegisterUserRequest
 ): Promise<
